requirements.txt
creditagricole-particuliers
Firefly-III-API-Client==2.0.12.0
python-dateutil==2.8.2